Break structural circular deps and tighten Fuji recipe typing
- Move LivePhotoVideoHandle into photo-viewer/types.ts to end the
  types.ts <-> LivePhotoVideo.tsx cycle; re-export from LivePhotoVideo
  for compat.
- Split shared map types into types/map/core.ts so provider.ts no longer
  loops back through the index barrel.
- Replace `as any` around processFujiRecipe with a mapped ProcessedFujiRecipe
  that widens only string fields, restoring type safety for ExifPanel.
- Drop `(window as any).__MANIFEST__` casts in photo detail page since
  Window is already augmented in global.d.ts.
- Drop the `as any` on slider.columns translation args.
- Harden applyAccentTransitionStyle: reuse an existing style element
  instead of stacking duplicate <style> tags under concurrent callers.
