+++ apps/web/src/pages/(main)/photos/[photoId]/index.tsx
@@ -26,7 +26,7 @@ export const Component = () => {
       console.warn('[PhotoDetail] Photos array is empty or not loaded yet', {
         photosLength: photos?.length || 0,
         photoId,
-        hasManifest: typeof window !== 'undefined' && (window as any).__MANIFEST__ !== undefined,
+        hasManifest: typeof window !== 'undefined' && window.__MANIFEST__ !== undefined,
       })
       return -1
     }
@@ -94,7 +94,7 @@ export const Component = () => {
       console.error('[PhotoDetail] Photo not found:', {
         requestedPhotoId: photoId,
         photosLength: photos?.length || 0,
-        hasManifest: (window as any).__MANIFEST__ !== undefined,
+        hasManifest: window.__MANIFEST__ !== undefined,
         photoIds: photos?.slice(0, 10).map((p) => p?.id) || [],
       })
     }
